[Refactor] Move HAsh module to src/common
