+++ .eslintrc.js
@@ -21,7 +21,7 @@ module.exports = {
     jest: true,
   },
   rules: {
-    'prettier/prettier': 'error',
+    "prettier/prettier": ["error", { "endOfLine": "auto" }],
     'import/extensions': [
       'error',
       'ignorePackages',
