+++ src/common/hash/hash.module.ts
