+++ src/common/hash/hash.service.ts
@@ -1,15 +1,14 @@
 import * as bcrypt from 'bcrypt';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import configuration from '../config/configuration';
+import configuration from '../../config/configuration';
 
 @Injectable()
 export class HashService {
   constructor(private readonly configService: ConfigService) {}
 
   async generateHash(password: string): Promise<string> {
-    const { salt } = configuration().hash;
-    return bcrypt.hash(password, salt);
+    return bcrypt.hash(password, configuration().saltRounds);
   }
 
   async compareHash(password: string, hash: string): Promise<boolean> {
