+++ src/users/user.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { HashModule } from '../hash/hash.module';
+import { HashModule } from '../common/hash/hash.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User]), HashModule],
